% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_list.R
\name{read_list}
\alias{read_list}
\title{Read in multiple files.}
\usage{
read_list(list_of_datasets, read_func, output = "list")
}
\arguments{
\item{list_of_datasets}{A list of file names with paths}

\item{read_func}{The import function}

\item{output}{Can be "list" or "global". If set to list the output is a list of objects, if "global" the output tables are created in the global namespace. Default is list.}
}
\value{
A list of, or multiple data.frames of the imported files.
}
\description{
Based on code from: http://www.brodrigues.co/2016/07/26/read-a-lot-of-datasets-at-once-with-r
}
\examples{
read_list(file_list, read.csv())
read_list(file_list, read.csv(), output = "list")
read_list(file_list, read.csv(), output = "global")
}

